+++ crates/teloxide-core/src/types/chat_member.rs
@@ -730,6 +730,7 @@ mod tests {
                 is_premium: false,
                 added_to_attachment_menu: false,
                 has_topics_enabled: false,
+                allows_users_to_create_topics: false,
             },
             kind: ChatMemberKind::Administrator(Administrator {
                 custom_title: None,
@@ -797,6 +798,7 @@ mod tests {
                 is_premium: false,
                 added_to_attachment_menu: false,
                 has_topics_enabled: false,
+                allows_users_to_create_topics: false,
             },
             kind: ChatMemberKind::Restricted(Restricted {
                 is_member: true,