+++ crates/teloxide-core/src/types/chat_owner_changed.rs
@@ -1,0 +1,13 @@
+use serde::{Deserialize, Serialize};
+
+use crate::types::User;
+
+/// Describes a service message about an ownership change in the chat.
+///
+/// [The official docs](https://core.telegram.org/bots/api#chatownerchanged).
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(test, derive(schemars::JsonSchema))]
+pub struct ChatOwnerChanged {
+    /// The new owner of the chat
+    pub new_owner: User,
+}