+++ crates/teloxide-core/src/types/transaction_partner.rs
@@ -255,6 +255,7 @@ mod tests {
                 is_premium: false,
                 added_to_attachment_menu: false,
                 has_topics_enabled: false,
+                allows_users_to_create_topics: false,
             },
             kind: TransactionPartnerUserKind::PremiumPurchase(
                 TransactionPartnerUserPremiumPurchase { premium_subscription_duration: Some(1) },