+++ crates/teloxide-core/src/types/unique_gift.rs
@@ -44,6 +44,11 @@ pub struct UniqueGift {
     #[serde(default, skip_serializing_if = "std::ops::Not::not")]
     pub is_premium: bool,
 
+    /// `true`, if the gift was used to craft another gift and isn't available
+    /// anymore
+    #[serde(default, skip_serializing_if = "std::ops::Not::not")]
+    pub is_burned: bool,
+
     /// Colors of the gift
     pub colors: Option<UniqueGiftColors>,
 }
@@ -64,6 +69,10 @@ pub struct UniqueGiftModel {
     /// The number of unique gifts that receive this model for every 1000 gifts
     /// upgraded
     pub rarity_per_mille: u32,
+
+    /// Rarity of the model if it is a crafted model. Currently, can be
+    /// "uncommon", "rare", "epic", or "legendary".
+    pub rarity: Option<String>,
 }
 
 /// This object describes the symbol shown on the pattern of a unique gift.
@@ -155,6 +164,7 @@ mod tests {
                 name: "name".to_owned(),
                 sticker: sticker.clone(),
                 rarity_per_mille: 123,
+                rarity: None,
             },
             symbol: UniqueGiftSymbol {
                 name: "name".to_owned(),
@@ -174,6 +184,7 @@ mod tests {
             publisher_chat: None,
             is_from_blockchain: false,
             is_premium: false,
+            is_burned: false,
             colors: None,
         };