+++ crates/teloxide-core/src/types/user_profile_audios.rs
@@ -1,0 +1,16 @@
+use serde::{Deserialize, Serialize};
+
+use crate::types::Audio;
+
+/// This object represents the audios displayed on a user's profile.
+///
+/// [The official docs](https://core.telegram.org/bots/api#userprofileaudios).
+#[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
+#[cfg_attr(test, derive(schemars::JsonSchema))]
+pub struct UserProfileAudios {
+    /// Total number of profile audios for the target user.
+    pub total_count: u32,
+
+    /// Requested profile audios.
+    pub audios: Vec<Audio>,
+}