+++ crates/teloxide/src/utils/html.rs
@@ -242,6 +242,7 @@ mod tests {
             is_premium: false,
             added_to_attachment_menu: false,
             has_topics_enabled: false,
+            allows_users_to_create_topics: false,
         };
         assert_eq!(user_mention_or_link(&user_with_username), "@abcd");
         let user_without_username = User {
@@ -254,6 +255,7 @@ mod tests {
             is_premium: false,
             added_to_attachment_menu: false,
             has_topics_enabled: false,
+            allows_users_to_create_topics: false,
         };
         assert_eq!(
             user_mention_or_link(&user_without_username),